requirements.txt
tensorflow
numpy
pillow
pycocotools
opencv-contrib-python